+++ bosen-product/src/main/java/com/bosen/product/vo/response/ProductDetailVO.java
@@ -68,7 +68,7 @@ public class ProductDetailVO {
     /**
      * 销量
      */
-    private Integer salesCounts;
+    private BigDecimal salesCounts;
 
     /**
      * 商品主图
